Service/requirements.txt
asgiref==3.4.1
paho-mqtt==1.5.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
